lambda_scraper/requirements.txt
requests
selectolax
psycopg2-binary